pyproject.toml
[tool.poetry]
name = "stepic-challenge"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
celery = "^5.1.2"
django-celery-results = "^2.2.0"
moviepy = "^1.0.3"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
